requirements.txt
pywinauto==0.6.8
pyautogui==0.9.54
pillow==10.4.0
mss==9.0.1
opencv-python==4.10.0.84
playwright==1.46.0
pydantic==2.8.2
pyyaml==6.0.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
numpy==2.1.0
